# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API built with Golang using Clean Code architecture",
        "title": "AnyFashion Application API",
        "contact": {
            "name": "API Support",
            "email": "abhinandarun11@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/clinic-create": {
            "post": {
                "description": "Create a new clinic with the specified attributes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new clinic",
                "parameters": [
                    {
                        "description": "Clinic details",
                        "name": "clinic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.Clinic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/api/create-ticket/{id}": {
            "post": {
                "description": "Create a new ticket for a specific clinic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Clinic ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token Number",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/api/get-ticket/{id}": {
            "get": {
                "description": "Display details of a specific ticket identified by its ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Show details of a specific ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "402": {
                        "description": "Error response",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "req.Clinic": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "res.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "stastuscode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerTokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}